Maps/sensor_locations/api/requirements.txt
flask
google-cloud-bigquery
pyproj
python-dateutil